+++ projects/authentication_management/zainab1AT/authRoutes.js
@@ -1,39 +1,23 @@
 import express from "express";
 import bcrypt from "bcrypt";
-import DAL from "./DAL.js";
 import { getConnection } from "./mysql.js";
+import DAL from "./DAL.js";
+import basicAuth from "./authMiddleware.js";
+import userService from "./userService.js";
+
 const router = express.Router();
 const con = getConnection();
 const DALInstance = new DAL(con);
-import basicAuth from "./authMiddleware.js";
 
-router.post("/signup",basicAuth, async (req, res) => {
-  const { username, password } = req.auth;
-  const { email, role } = req.body;
-  const hashedPassword = await bcrypt.hash(password, 12);
-  try {
-    const user = await DALInstance.findByUsername(username);
-    if (!user) {
-      await DALInstance.addUser(username, email, hashedPassword, role);
-      res.status(201).json({ message: "User registered successfully" });
-    } else res.status(500).json({ message: "Registration faild" });
-  } catch (error) {
-    res.status(500).json({ error: "Registration failed" });
-  }
-});
 
-router.post("/login",basicAuth, async (req, res) => {
+router.get("/all",basicAuth, async (req, res) => {
   const { username, password } = req.auth;
+
   try {
-    const user = await DALInstance.findByUsername(username);
-    if (user) {
-      const match = await bcrypt.compare(password, user.password);
-      if (match)
-        res.status(200).json({ message: "User logged in successfully" });
-      else res.status(401).json({ message: "Invalid username or password" });
-    } else res.status(401).json({ message: "Invalid username or password" });
+  const users = await userService.allUsers(username, password);
+    res.status(200).json({ users: users });
   } catch (error) {
-    res.status(500).json({ error: "Login failed" });
+    res.status(401).json({ message: error.message });
   }
 });
